Bug fixes for handling dynamic updates within a For component.
-) IComponent.peekClientId() / FormBehavior.peekClientId(IFormComponent) methods added so that 
DojoAjaxResponseBuilder can properly determine if a component was requested to be updated based on the client 
id that it ~will~ render with. 

-) Core javascript html fragment updating logic changed to apply style/class attributes to elements being updated.

-) For component refactored to properly interact with ComponentRenderWorker chain on each iteration of render 
as well as the generateClientId semantics. This area still needs more refactoring as found in the TODO comments 
of the ForBean class.

-) AbstractLinkComponent and associated areas re-factored so that the clientId attribute is written ~before~ generating
the url so that the clientId is correctly known (and generated on the fly) when needed. Again, another area
that smells of needing refactoring. Things are becoming slightly too fragile if people have to think about 
these kinds of semantics too much.


git-svn-id: https://svn.apache.org/repos/asf/tapestry/tapestry4/trunk@479189 13f79535-47bb-0310-9956-ffa450edef68