Fixes TAPESTRY-1175, pending further testing on windows to be sure.
The default unprotected resource regexp paths allowed for access to non tapestry resources if the project
included a string of "tapestry" or "dojo" in their package name. (a very likely scenerio) . Fix was to add a
begin of line character to the unprotected paths. The result was that specific new regexp's had to be added 
to handle all of the default css/images/etc bundled with tapestry as those packages were now being correctly 
excluded. The begin of line fix was suggested on the dev list by Bruce Snyder.

git-svn-id: https://svn.eu.apache.org/repos/asf/tapestry/tapestry4/trunk@483259 13f79535-47bb-0310-9956-ffa450edef68