Resolves TAPESTRY-1253.
First was that injections of type meta with no previous "java property" equivalent will have a default type of Object, but this property is read only since you're using the very basic meta property type. (ie read only)

Added a new "meta:" binding that will hopefully be used in 97% of the cases like yours. It works in any sort of property syntax expression so it would be easy to specify it as the intitial-value to a normal property.

Updated documentation.


git-svn-id: https://svn.apache.org/repos/asf/tapestry/tapestry4/trunk@508289 13f79535-47bb-0310-9956-ffa450edef68