Fixes TAPESTRY-1513.  Added a new configuration type for BindingSource which controls which binding factory to use based on the name of the property the binding is going to be bound to - if known.  The logic only kicks in when no binding prefix has been specified.
The point in adding it was so that people can use statements like updateComponents="compA, compB" and have the system automatically resolve the unique clientId for each component specified depending on the context in which the binding is invoked.  Sort of like the clientid: binding only this one kicks in automatically and may be easier for people to understand than throwing another binding type at them.

git-svn-id: https://svn.eu.apache.org/repos/asf/tapestry/tapestry4/trunk@542045 13f79535-47bb-0310-9956-ffa450edef68